Scalars: warnings begone
except, negative integer implicitly converted to unsigned ....

+++ src/Engine/Scalers/common.h
@@ -47,10 +47,11 @@ static inline uint32_t rgb_to_yuv(uint32_t c)
 }
 
 /* Test if there is difference in color */
-static inline int yuv_diff(uint32_t yuv1, uint32_t yuv2) {
-	return (( abs((int)((yuv1 & Ymask) - (yuv2 & Ymask))) > trY ) ||
-			( abs((int)((yuv1 & Umask) - (yuv2 & Umask))) > trU ) ||
-			( abs((int)((yuv1 & Vmask) - (yuv2 & Vmask))) > trV ) );
+static inline int yuv_diff(uint32_t yuv1, uint32_t yuv2)
+{
+	return (( abs(static_cast<int>((yuv1 & Ymask) - (yuv2 & Ymask))) > trY ) ||
+			( abs(static_cast<int>((yuv1 & Umask) - (yuv2 & Umask))) > trU ) ||
+			( abs(static_cast<int>((yuv1 & Vmask) - (yuv2 & Vmask))) > trV ) );
 }
 
 static inline int Diff(uint32_t c1, uint32_t c2)
@@ -66,17 +67,17 @@ static inline uint32_t Interpolate_2(uint32_t c1, int w1, uint32_t c2, int w2, i
 		return c1;
 	}
 	return
-		(((((c1 & MASK_ALPHA) >> 24) * w1 + ((c2 & MASK_ALPHA) >> 24) * w2) << (24-s)) & MASK_ALPHA) +
-		((((c1 & MASK_2) * w1 + (c2 & MASK_2) * w2) >> s) & MASK_2)	+
-		((((c1 & MASK_13) * w1 + (c2 & MASK_13) * w2) >> s) & MASK_13);
+		(((((c1 & MASK_ALPHA) >> 24) * static_cast<uint32_t>(w1) + ((c2 & MASK_ALPHA) >> 24) * static_cast<uint32_t>(w2)) << (24-s)) & MASK_ALPHA) +
+		((( (c1 & MASK_2)            * static_cast<uint32_t>(w1) +  (c2 & MASK_2)            * static_cast<uint32_t>(w2)) >> s)      & MASK_2)	+
+		((( (c1 & MASK_13)           * static_cast<uint32_t>(w1) +  (c2 & MASK_13)           * static_cast<uint32_t>(w2)) >> s)      & MASK_13);
 }
 
 static inline uint32_t Interpolate_3(uint32_t c1, int w1, uint32_t c2, int w2, uint32_t c3, int w3, int s)
 {
 	return
-		(((((c1 & MASK_ALPHA) >> 24) * w1 + ((c2 & MASK_ALPHA) >> 24) * w2 + ((c3 & MASK_ALPHA) >> 24) * w3) << (24-s)) & MASK_ALPHA) +
-		((((c1 & MASK_2) * w1 + (c2 & MASK_2) * w2 + (c3 & MASK_2) * w3) >> s) & MASK_2) +
-		((((c1 & MASK_13) * w1 + (c2 & MASK_13) * w2 + (c3 & MASK_13) * w3) >> s) & MASK_13);
+		(((((c1 & MASK_ALPHA) >> 24) * static_cast<uint32_t>(w1) + ((c2 & MASK_ALPHA) >> 24) * static_cast<uint32_t>(w2) + ((c3 & MASK_ALPHA) >> 24) * static_cast<uint32_t>(w3)) << (24-s)) & MASK_ALPHA) +
+		((( (c1 & MASK_2)            * static_cast<uint32_t>(w1) +  (c2 & MASK_2)            * static_cast<uint32_t>(w2) +  (c3 & MASK_2)            * static_cast<uint32_t>(w3)) >> s)      & MASK_2) +
+		((( (c1 & MASK_13)           * static_cast<uint32_t>(w1) +  (c2 & MASK_13)           * static_cast<uint32_t>(w2) +  (c3 & MASK_13)           * static_cast<uint32_t>(w3)) >> s)      & MASK_13);
 }
 
 static inline uint32_t Interp1(uint32_t c1, uint32_t c2)
